requirements.txt
blinker==1.5
eventlet==0.33.1
Flask-APScheduler==1.12.4
Flask-SocketIO==5.3.0
flask-sqlalchemy==2.5.1
Flask-MQTT==1.1.1
